edit homework-1 #2
