+++ 2016/TalehZaliyev/1/calc_rpn.rb
@@ -1,16 +1,9 @@
 #!/usr/bin/env ruby
-class Numeric
-  def bit_rep(num_bit)
-    a = self - self.to_i
-    self.to_i.to_s(2).split('').reverse.map do |x|
-      (x == "1" && num_bit>0) ? ((num_bit -= 1) && (x = "0")) : x
-    end.reverse.join.to_i(2) + a
-  end
-end
+require './lib/Numeric.rb'
 def calc_rpn(stack = [])
-  while(true)
-    case(input = gets.strip)
-    when %r{^(\d+|\d+[.]\d+)$}; stack.push(input)
+  loop do
+    case (input = gets.strip)
+    when /^(\d+|\d+[.]\d+)$/ then stack.push(input)
     when %r{^[-+*\/!]$}
       input == "!" ? input = ".bit_rep " + stack.pop : input += " " + stack.pop
       stack.push(eval(stack.pop + ".to_f" + input).to_s)
